extract compute computeSelector
