+++ test/stylesheet-registry.js
@@ -5,6 +5,7 @@ import test from 'ava'
 import { StyleSheetRegistry } from '../src/stylesheet-registry'
 import makeSheet, { invalidRules } from './stylesheet'
 import withMock, { withMockDocument } from './helpers/with-mock'
+import { computeId, computeSelector } from '../src/lib/hash'
 
 function makeRegistry(options = { optimizeForSpeed: true, isBrowser: true }) {
   const registry = new StyleSheetRegistry({
@@ -254,8 +255,8 @@ test(
 test(
   'createComputeId',
   withMock(withMockDocument, t => {
-    const utilRegistry = makeRegistry()
-    const computeId = utilRegistry.createComputeId()
+    // const utilRegistry = makeRegistry()
+    // const computeId = utilRegistry.createComputeId()
 
     // without props
     t.is(computeId('123'), 'jsx-123')
@@ -270,10 +271,10 @@ test(
 test(
   'createComputeSelector',
   withMock(withMockDocument, t => {
-    const utilRegistry = makeRegistry()
-    const computeSelector = utilRegistry
-      .createComputeSelector()
-      .bind(utilRegistry)
+    // const utilRegistry = makeRegistry()
+    // const computeSelector = utilRegistry
+    // .createComputeSelector()
+    // .bind(utilRegistry)
 
     t.is(
       computeSelector(
